feat(prepare-commit-msg): introduce SKIP_PREPARE_COMMIT_MSG
Prior to this change, the only way to avoid Anthropic API usage was to
remove the secret.

This change introduces an environment variable, namely
SKIP_PREPARE_COMMIT_MSG, to skip the call to Anthropic's API.  Also, the
unit price is baked into the tool and usage cost displayed with the
token counts.

Signed-off-by: Corin Lawson <corin@phiware.com.au>