feat: Introduce Versent’s SOC2 requirements
This change introduces Versent’s SOC2 security requirements into the
NixOS configuration. Specifically, it includes:

1. Enabling the CrowdStrike Falcon Sensor as a systemd service. The
   Falcon Sensor provides endpoint security and monitoring capabilities.
2. Enabling the Cloudflare WARP service and including the Cloudflare CA
   certificate. Cloudflare WARP provides secure internet connectivity
   and protects against network-based threats.

The SOC2 requirements have been modularized into separate Nix modules
that can be easily enabled or disabled as needed. This allows for a more
flexible and maintainable configuration.

Additionally, some unused packages have been removed from the
`euler.nix` file to keep the overall package set clean and focused.

Note that crowdstrike module requires the CID to be specified, I'm not
sure if this is sensitive information or not so I have not committed in
this change.  The source of this parameter may change in the future.
Versent provide this value at step 2 of
[Install Versent SOC2 Controls on Linux](https://versent.atlassian.net/wiki/spaces/CYBER/pages/8974073979/Install+Versent+SOC2+Controls+on+Linux#CrowdStrike).

Signed-off-by: Corin Lawson <corin@phiware.com.au>