feat: include branch name in commit message generation
Prior to this change, the commit message generation script was only
taking the diff content into account when making the API request to the
language model. This change adds the current branch name as an
additional input, making the generated commit messages more
context-aware and able to reference a ticket or issue number if it is
used in the branch name.

The branch name is now passed as an argument to the `makeAPICall`
function, and the prompt template has been updated to include the branch
name as an input variable. The tests have also been updated to verify
the inclusion of the branch name in the API request content.

This change should help produce more informative and relevant commit
messages that adhere to the conventional commits style.