Introduce lix to replace nix
Prior to this change, I was experiencing "some curious behavior" when
entering the dev shell for paas-payment-initiation. "Specifically it
spends some significant amount of time copying a store-path into the
store."[1]

This change replaces the nix package with lix in an attempt to avoid the
copying of a store path.

[1]: https://discourse.nixos.org/t/nix-copying-a-store-path-into-the-store/60409

Signed-off-by: Corin Lawson <Corin.Lawson+cba@cba.com.au>