Introduce example flake
Prior to this change, I just wanted to share the flake that I use for
day-to-day development.

This change includes the flake that I use for the Typescript repo that I
use daily. It is intended to be copied to a directory that then has a
number of subdirectories that contain separate work trees fo a single
repo. This reduces the churn that I see when moving between directories
as there is a long standing, significant [bug] in nix.

This file is intended to be used with direnv and a `.envrc` file that
simply contains `use flake .`. Additionally, consider adding exports to
change frequently to `.envrc` to avoid rebuilding the flake.

[bug]: https://github.com/NixOS/nix/issues/11098

Signed-off-by: Corin Lawson <corin@phiware.com.au>